assign01/book_store_dnl/Microservices Django/cart-service/requirements.txt
Django>=5.2.10
djangorestframework>=3.14.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
